pyproject.toml
[project]
name = "rd-achievements-exhibition"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openpyxl>=3.1.0",
    "pandas>=2.0.0",
    "pillow>=10.0.0",
    "pypdf2>=3.0.0",
    "qrcode[pil]>=7.4.0",
    "rapidfuzz>=3.14.3",
    "reportlab>=4.0.0",
]
